requirements.txt
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
icalendar==5.0.11